[#165797252] Add setup to only use DB migrations in test
Because the regional DAHLIA listings site and admin site are sharing a database, we are consolidating migrations into just one app to keep things simple for now. We are running migrations from the admin app, so we will not be keeping migrations in the listings app. For running tests on the listings app, we will generate the migrations from the dahlia_data_models gem for the purposes of creating the test DB, and then delete the migrations when the tests are done.