Add e2e tests for GroupSwitchable and Angular views #165797252 (#12)
* [#164584674] Integrate group model updates

* [#165797870] Switch from npm to yarn and upgrade node version

Also update command for updating webdriver for protractor to use an updated syntax that can pull in the latest driver versions.

* [#165797870] Check in yarn package resolutions

* [#165797252] Add e2e tests for GroupSwitchable and Angular views

* [#165797252] Add setup to only use DB migrations in test

Because the regional DAHLIA listings site and admin site are sharing a database, we are consolidating migrations into just one app to keep things simple for now. We are running migrations from the admin app, so we will not be keeping migrations in the listings app. For running tests on the listings app, we will generate the migrations from the dahlia_data_models gem for the purposes of creating the test DB, and then delete the migrations when the tests are done.

* [#165797252] Add bundle exec to rails command for Semaphore

* [#165797252] Make header logo group class name more specific