[#165892070] Remove hardcoded cases for specific SF listings
In SF DAHLIA, there were certain specific listings that had requirements unique to just that listing, so there were some hardcoded cases in the Angular app specific to those listings. We of course don't need any of that on regional DAHLIA because those specific listings won't be on regional DAHLIA, so we are removing the hardcoding.