[#165979384] Return to the version of ui-router we had been using
The newer version of ui-router was changing the behavior when a user attempted to view a listing page using an invalid ID. The previous, expected behavior was for the user to be redirected. The new behavior was that the Angular app showed an error alert. We will determine later when/how to deal with upgrading ui-router and refactoring our use of it. We may end up moving on from AngularJS before we end up doing this.