Move from bower to yarn and upgrade Angular to 1.7.8 #165979384 (#30)
* [#165979384] First pass at moving from Bower to Yarn

* [#165979384] Stable build now with Angular 1.7.8

* [#165979384] Fix Jasmine tests and remove http-etag module

* [#165979384] Add Yarn for Heroku support

* [#165979384] Remove yarn.lock from .slugignore, remove httpEtagProvider

* [#165979384] Remove error intercept from API controller

This interception was tailored for Salesforce errors. It's no longer relevant to the current state of regional DAHLIA.

* [#165979384] Update component controllers to use $onInit

Use of $onInit for referencing bindings is required since AngularJS 1.6: https://docs.angularjs.org/guide/migration#migrating-from-1-5-to-1-6. And using $onInit for everything that's not a public method is recommended even if bindings are not used.

* [#165979384] Remove references to bower

* [#165979384] Package updates

Remove angular-http-etag from package.json because we're no longer using it. Specify version for eslint's is-my-json-valid dependency to avoid vuln in an earlier version that we were using.

* [#165979384] Return to the version of ui-router we had been using

The newer version of ui-router was changing the behavior when a user attempted to view a listing page using an invalid ID. The previous, expected behavior was for the user to be redirected. The new behavior was that the Angular app showed an error alert. We will determine later when/how to deal with upgrading ui-router and refactoring our use of it. We may end up moving on from AngularJS before we end up doing this.

* [#165979384] Update to latest version of dahlia_data_models gem

* [#165979384] Update listingContainer jasmine test to test $onInit