[#167584018] Convert monthly rent to int in unit summary (#59)
The monthly rent is a decimal in the DB, so when it gets converted to JSON it gets represented as a string instead of a number. This was causing problems with our display of the unit monthly rent range on the listings page. The rents on all current SMC and SJ units are integers, and we never display a monthly rent figure as a float in the listings site frontend, so doing integer conversion of the monthly rents before sending to the frontend is fine. We should revisit all listing field types in our renovation work.